+++ CHANGELOG.md
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added caption to Table
 
 ### Changed
+- Improved accessibility of the input's character counter
+
+### Changed
 
 - Core branding update 6.6.4